fix in nodemon local and rename files
