+++ index.js
@@ -1,4 +1,4 @@
-require('babel-core/register');
+require('babel-register');
 require('babel-polyfill');
 require('css-modules-require-hook');
 require('./server/server');
