+++ nodemon.json
@@ -1,0 +1,13 @@
+{
+  "ignore": [
+    "node_modules/**/node_modules"
+  ],
+  "verbose": true,
+  "watch": [
+    "./server/"
+  ],
+  "env": {
+    "NODE_ENV": "development"
+  },
+  "ext": "js"
+}
