+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "test": "mocha shared/tests/*.spec.js --compilers js:babel-register",
     "test:server": "cross-env NODE_ENV=test PORT=8080 MONGO_URL=mongodb://localhost:27017/mern-test mocha --compilers js:babel-register --recursive server/tests/**/*.spec.js",
-    "start": "cross-env NODE_ENV=development nodemon index.js --watch server",
+    "start": "cross-env NODE_ENV=development nodemon index.js",
     "start:prod": "cross-env NODE_ENV=production node index.js",
     "bs": "npm run clean && npm run build && npm run start:prod",
     "minify": "cleancss -o static/css/app.min.css static/css/app.css",
@@ -55,7 +55,7 @@
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "babel-preset-react-hmre": "^1.1.0",
-    "babel-register": "^6.4.3",
+    "babel-register": "^6.6.5",
     "chai": "^3.5.0",
     "clean-css": "^3.4.9",
     "cross-env": "^1.0.7",
