+++ shared/redux/reducers/index.js
@@ -1,9 +1,9 @@
-import blog from './blog';
+import post from './post';
 import { routerReducer as routing } from 'react-router-redux';
 import { combineReducers } from 'redux';
 
 const rootReducer = combineReducers({
-  blog,
+  post,
   routing,
 });
